refactor(enrich): extract enrichAndFindContacts service
Move enrichOrganization, getActiveSignalSlugs, the SIGNAL_SLUG_*
constants, and the findContactsForCompany wrapper out of the
enrich-company route into src/lib/services/company-enrichment.ts.
The route keeps auth, link-id-vs-org-id resolution, withAction cost
attribution, and its exact response shapes; the service returns plain
data.

Every moved function now takes an explicit SupabaseClient, and the
helpers in its call chain that built their own RLS client internally
(isRecentlyEnriched, mergeEnrichmentData, findOrCreatePerson,
linkPersonToCampaign) accept an optional client that contact-discovery
and the new service thread through — so the whole chain works under
the QStash admin client later. Existing callers are unchanged (the
optional param defaults to the request RLS client).

No behavior change: the 7-day recency skip still runs contact-finding,
and trackUsage operation names are untouched.
